requirements/base.txt
#Django
django==2.1  # https://www.djangoproject.com/

Pillow==5.2.0  # https://github.com/python-pillow/Pillow

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
djangorestframework-jwt
markdown==2.1
coreapi==2.3.3
django-rest-swagger==2.2.0
django-cors-headers==2.4.0
rest-framework-generic-relations==1.2.1
django-filter==2
django-polymorphic==2.0.3


gunicorn
#SQLAlchemy==1.2.8
mysql
mock
django-redis
celery==4.2.1


pytz==2018.9
nepali==0.3.0
django-import-export

scipy
